llc_sweeper/src/llc_sweeper/magnetics/openmagnetics_adapter.py

"""
Adapter for OpenMagnetics Integration.
Handles optional dependency gracefully.
"""

import sys
from typing import Optional, Dict, Any, List
import logging

# Configure logger
logger = logging.getLogger(__name__)

# Try to import PyOpenMagnetics
try:
    # Hypothetical imports - adjust if package structure differs
    import openmagnetics
    # from openmagnetics import DesignAdviser ? 
    OPENMAGNETICS_AVAILABLE = True
except ImportError:
    OPENMAGNETICS_AVAILABLE = False


def is_openmagnetics_available() -> bool:
    return OPENMAGNETICS_AVAILABLE


def design_transformer_openmagnetics(
    specs: Any, # Typed as LLCSpecs but avoiding circle import if possible, or use explicit type
    res: Any,   # Typed as SimulationResult
    corner: str = "full_load"
) -> Dict[str, Any]:
    """
    Propose transformer designs using OpenMagnetics.
    
    Inputs:
    - n_used (Turn ratio)
    - Lm_real (Magnetizing Inductance)
    - fsw (Switching Frequency)
    - Waveforms (Approximated)
    """
    
    result = {
        "status": "fail",
        "top_designs": [],
        "chosen": None,
        "warnings": [],
        "errors": [],
        "metrics": {
             "total_loss_W": 0.0,
             "core_loss_W": 0.0,
             "copper_loss_W": 0.0,
             "estimated_temp_rise_C": 0.0,
             "fill_factor": 0.0,
             "volume_cm3": 0.0
        }
    }

    if not OPENMAGNETICS_AVAILABLE:
        result["errors"].append("OpenMagnetics not installed.")
        return result

    try:
        # Placeholder for Real Logic:
        # 1. Define Core requirements (Power, Freq, Lm)
        # 2. Define Winding requirements (Turns Ratio, Current)
        # 3. Call Design Adviser
        
        # Example pseudo-code for when library is available:
        # mag = openmagnetics.MagneticComponent(...)
        # adviser = openmagnetics.DesignAdviser()
        # designs = adviser.suggest(mag)
        
        # For now, since we don't have the library to test, we return fail 
        # but with a note that it's a stub.
        result["errors"].append("OpenMagnetics logic not yet fully mapped (Stub).")
        result["status"] = "fail"
        
        # If successful:
        # result["status"] = "ok"
        # result["top_designs"] = [ ... ]
        # result["chosen"] = ...
        
    except Exception as e:
        logger.error(f"Error in design_transformer_openmagnetics: {e}")
        result["errors"].append(str(e))
        result["status"] = "fail"
        
    return result


def design_resonant_inductor_openmagnetics(
    specs: Any,
    res: Any,
    corner: str = "full_load"
) -> Dict[str, Any]:
    """
    Propose resonant inductor designs using OpenMagnetics.
    
    Inputs:
    - Lr_real (Resonant Inductance)
    - fsw
    - Current (RMS/Peak)
    """
    result = {
        "status": "fail",
        "top_designs": [],
        "chosen": None,
        "warnings": [],
        "errors": [],
        "metrics": {
             "total_loss_W": 0.0,
             "core_loss_W": 0.0,
             "copper_loss_W": 0.0,
             "estimated_temp_rise_C": 0.0,
             "fill_factor": 0.0,
             "volume_cm3": 0.0
        }
    }

    if not OPENMAGNETICS_AVAILABLE:
        result["errors"].append("OpenMagnetics not installed.")
        return result

    try:
        # Placeholder for Inductor Logic
        result["errors"].append("OpenMagnetics logic not yet fully mapped (Stub).")
        result["status"] = "fail"
        
    except Exception as e:
        logger.error(f"Error in design_resonant_inductor_openmagnetics: {e}")
        result["errors"].append(str(e))
        result["status"] = "fail"

    return result
