llc_sweeper/requirements.txt
streamlit
pandas
matplotlib
numpy
scipy
PyOpenMagnetics
